moved some components around
