+++ src/client/Types/taskDataType.ts
@@ -1,0 +1,7 @@
+/* eslint-disable semi */
+export default interface taskData {
+  name: string,
+  lastCompleted: string,
+  frequency: string,
+  id: number,
+}
